+++ api-surface/vigiles-testing.api.md
@@ -528,22 +528,10 @@ export interface HookPropertyResult<E> {
 }
 
 // @public (undocumented)
-export interface HookRunResult {
+export interface HookRunResult extends ScriptRunResult {
     readonly blocked: boolean;
     readonly decision: HookOutput["decision"] | "allow" | "deny" | "ask" | undefined;
-    readonly egress: readonly EgressAttempt[];
-    readonly egressDropped?: {
-        readonly packets: number;
-        readonly bytes: number;
-    };
-    // (undocumented)
-    readonly exitCode: number;
-    readonly filesWritten: readonly string[];
     readonly json: HookOutput | null;
-    // (undocumented)
-    readonly stderr: string;
-    // (undocumented)
-    readonly stdout: string;
 }
 
 // @public
@@ -768,18 +756,8 @@ export interface RunHarnessTestOptions {
 // @public
 export function runHook(command: string, input: HookInput, opts?: RunHookOptions): HookRunResult;
 
-// @public (undocumented)
-export interface RunHookOptions {
-    readonly cwd?: string;
-    readonly egress?: {
-        readonly allow: readonly string[];
-    };
-    readonly env?: Record<string, string>;
-    readonly recordEgress?: boolean;
-    readonly sandbox?: SandboxMode;
-    readonly timeoutMs?: number;
-    readonly trusted?: boolean;
-}
+// @public
+export type RunHookOptions = Omit<RunScriptOptions, "stdin">;
 
 // @public
 export interface RunOut {
@@ -791,7 +769,38 @@ export interface RunOut {
 }
 
 // @public
+export function runScript(command: string, opts?: RunScriptOptions): ScriptRunResult;
+
+// @public (undocumented)
+export interface RunScriptOptions {
+    readonly cwd?: string;
+    readonly egress?: {
+        readonly allow: readonly string[];
+    };
+    readonly env?: Record<string, string>;
+    readonly recordEgress?: boolean;
+    readonly sandbox?: SandboxMode;
+    readonly stdin?: string;
+    readonly timeoutMs?: number;
+    readonly trusted?: boolean;
+}
+
+// @public
 export type SandboxMode = "auto" | "strict" | false;
+
+// @public
+export interface ScriptRunResult {
+    readonly egress: readonly EgressAttempt[];
+    readonly egressDropped?: {
+        readonly packets: number;
+        readonly bytes: number;
+    };
+    readonly exitCode: number;
+    readonly filesWritten?: readonly string[];
+    readonly stderr: string;
+    // (undocumented)
+    readonly stdout: string;
+}
 
 // @public
 export interface SelectionTrialResult {
