+++ api-surface/vigiles-unit.api.md
@@ -314,22 +314,10 @@ export interface HookPropertyResult<E> {
 }
 
 // @public (undocumented)
-export interface HookRunResult {
+export interface HookRunResult extends ScriptRunResult {
     readonly blocked: boolean;
     readonly decision: HookOutput["decision"] | "allow" | "deny" | "ask" | undefined;
-    readonly egress: readonly EgressAttempt[];
-    readonly egressDropped?: {
-        readonly packets: number;
-        readonly bytes: number;
-    };
-    // (undocumented)
-    readonly exitCode: number;
-    readonly filesWritten: readonly string[];
     readonly json: HookOutput | null;
-    // (undocumented)
-    readonly stderr: string;
-    // (undocumented)
-    readonly stdout: string;
 }
 
 // @public
@@ -424,8 +412,14 @@ export function requestContains(trace: Trace, needle: string | RegExp): boolean;
 // @public
 export function runHook(command: string, input: HookInput, opts?: RunHookOptions): HookRunResult;
 
+// @public
+export type RunHookOptions = Omit<RunScriptOptions, "stdin">;
+
+// @public
+export function runScript(command: string, opts?: RunScriptOptions): ScriptRunResult;
+
 // @public (undocumented)
-export interface RunHookOptions {
+export interface RunScriptOptions {
     readonly cwd?: string;
     readonly egress?: {
         readonly allow: readonly string[];
@@ -433,8 +427,23 @@ export interface RunHookOptions {
     readonly env?: Record<string, string>;
     readonly recordEgress?: boolean;
     readonly sandbox?: SandboxMode;
+    readonly stdin?: string;
     readonly timeoutMs?: number;
     readonly trusted?: boolean;
+}
+
+// @public
+export interface ScriptRunResult {
+    readonly egress: readonly EgressAttempt[];
+    readonly egressDropped?: {
+        readonly packets: number;
+        readonly bytes: number;
+    };
+    readonly exitCode: number;
+    readonly filesWritten?: readonly string[];
+    readonly stderr: string;
+    // (undocumented)
+    readonly stdout: string;
 }
 
 // @public
